+++ app/auth/login.css
@@ -1,0 +1,6 @@
+body {
+  margin: 0;
+  overflow: hidden;
+  font-family: var(--fonts-sans);
+  font-size: var(--fontSizes-1);
+}
