+++ app/routes/logout.tsx
@@ -2,7 +2,7 @@ import { ActionFunction } from "@remix-run/node";
 import { authenticator } from "~/services/auth.server";
 
 export default function Login() {
-  return "loging you out";
+  return "";
 }
 
 export const loader: ActionFunction = async ({ request }) => {
